src/test/java/com/uniprojecao/fabrica/gprojuridico/services/UserServiceTest.java
package com.uniprojecao.fabrica.gprojuridico.services;

import com.uniprojecao.fabrica.gprojuridico.UsuarioAggregator;
import com.uniprojecao.fabrica.gprojuridico.domains.usuario.Usuario;
import com.uniprojecao.fabrica.gprojuridico.repository.UserRepository;
import org.junit.jupiter.api.Nested;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.aggregator.AggregateWith;
import org.junit.jupiter.params.provider.CsvFileSource;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.mock;

@ExtendWith(MockitoExtension.class)
class UserServiceTest {

    @Mock
    private UserRepository repository;
    @InjectMocks
    private UserService service;

    @ParameterizedTest
    @CsvFileSource(resources = "/usuarios.csv")
    void create(@AggregateWith(UsuarioAggregator.class) Usuario usuario) {
        var userCreated = service.create(usuario);
        assertNotNull(userCreated.get("id"));
    }

    @Test
    void findById() {}

    @Test
    void update() {}

    @Test
    void delete() {}

    @Nested
    class Exception {}
}